# openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "wearable-analytics-api",
    "version": "0.1.0",
    "description": "OpenAPI specification for the wearable analytics API service.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development server"
    }
  ],
  "security": [],
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "issues": {
            "type": "object",
            "properties": {
              "formErrors": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "fieldErrors": {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "required": [
              "formErrors",
              "fieldErrors"
            ]
          }
        },
        "required": [
          "error",
          "issues"
        ]
      },
      "SyncErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "reauthenticationRequired": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "OuraConnectResponse": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "oura"
            ]
          },
          "authorizationUrl": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "provider",
          "authorizationUrl"
        ]
      },
      "OuraConnectUnavailableResponse": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "oura"
            ]
          },
          "configured": {
            "type": "boolean"
          },
          "missingConfiguration": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "requestedScopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "connected": {
            "type": "boolean"
          },
          "needsReconnect": {
            "type": "boolean"
          },
          "connection": {
            "type": "object",
            "nullable": true,
            "properties": {
              "isActive": {
                "type": "boolean"
              },
              "connectedAt": {
                "type": "string",
                "format": "date-time"
              },
              "expiresAt": {
                "type": "string",
                "format": "date-time"
              },
              "scopes": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "tokenExpired": {
                "type": "boolean"
              }
            },
            "required": [
              "isActive",
              "connectedAt",
              "expiresAt",
              "scopes",
              "tokenExpired"
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "provider",
          "configured",
          "missingConfiguration",
          "requestedScopes",
          "connected",
          "needsReconnect",
          "connection",
          "message"
        ]
      },
      "OuraDisconnectResponse": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "oura"
            ]
          },
          "disconnected": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "connected": {
            "type": "boolean",
            "enum": [
              false
            ]
          }
        },
        "required": [
          "provider",
          "disconnected",
          "connected"
        ]
      },
      "RunSyncResponse": {
        "type": "object",
        "properties": {
          "syncRun": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "userId": {
                "type": "string"
              },
              "source": {
                "type": "string",
                "enum": [
                  "oura"
                ]
              },
              "mode": {
                "type": "string",
                "enum": [
                  "manual",
                  "scheduled",
                  "backfill"
                ]
              },
              "status": {
                "type": "string",
                "enum": [
                  "pending",
                  "running",
                  "succeeded",
                  "failed"
                ]
              },
              "rangeStart": {
                "type": "string",
                "nullable": true,
                "format": "date-time"
              },
              "rangeEnd": {
                "type": "string",
                "nullable": true,
                "format": "date-time"
              },
              "startedAt": {
                "type": "string",
                "nullable": true,
                "format": "date-time"
              },
              "finishedAt": {
                "type": "string",
                "nullable": true,
                "format": "date-time"
              },
              "errorMessage": {
                "type": "string",
                "nullable": true
              },
              "createdAt": {
                "type": "string",
                "format": "date-time"
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time"
              }
            },
            "required": [
              "id",
              "userId",
              "source",
              "mode",
              "status",
              "rangeStart",
              "rangeEnd",
              "startedAt",
              "finishedAt",
              "errorMessage",
              "createdAt",
              "updatedAt"
            ]
          },
          "summary": {
            "type": "object",
            "properties": {
              "sleepRecords": {
                "type": "integer"
              },
              "readinessRecords": {
                "type": "integer"
              },
              "activityRecords": {
                "type": "integer"
              }
            },
            "required": [
              "sleepRecords",
              "readinessRecords",
              "activityRecords"
            ]
          },
          "window": {
            "type": "object",
            "properties": {
              "startDate": {
                "type": "string",
                "format": "date"
              },
              "endDate": {
                "type": "string",
                "format": "date"
              }
            },
            "required": [
              "startDate",
              "endDate"
            ]
          }
        },
        "required": [
          "syncRun",
          "summary",
          "window"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/integrations/oura/connect": {
      "post": {
        "tags": [
          "Oura Integration"
        ],
        "operationId": "createOuraConnectUrl",
        "summary": "Create a new Oura authorization URL",
        "responses": {
          "201": {
            "description": "Created a fresh Oura authorization URL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OuraConnectResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many connect attempts.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Oura OAuth is not configured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OuraConnectUnavailableResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/integrations/oura/callback": {
      "get": {
        "tags": [
          "Oura Integration"
        ],
        "operationId": "handleOuraCallback",
        "summary": "Handle the Oura OAuth callback",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": false,
            "name": "code",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": false,
            "name": "error",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "scope",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": false,
            "name": "state",
            "in": "query"
          }
        ],
        "responses": {
          "302": {
            "description": "Redirects back to the web app after the callback is handled.",
            "headers": {
              "Location": {
                "description": "Redirect target.",
                "schema": {
                  "type": "string",
                  "format": "uri"
                }
              }
            }
          },
          "429": {
            "description": "Too many callback attempts.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/integrations/oura/disconnect": {
      "post": {
        "tags": [
          "Oura Integration"
        ],
        "operationId": "disconnectOura",
        "summary": "Disconnect the local Oura connection",
        "responses": {
          "200": {
            "description": "Disconnected the local Oura connection.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OuraDisconnectResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many disconnect attempts.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/oura/run": {
      "post": {
        "tags": [
          "Sync"
        ],
        "operationId": "runOuraSync",
        "summary": "Run a manual Oura sync",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "startDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "lookbackDays": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true,
                    "maximum": 365
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Started and completed a manual Oura sync.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSyncResponse"
                }
              }
            }
          },
          "400": {
            "description": "The sync request payload is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Oura reauthentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "An Oura sync is already running.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "The sync failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/oura/backfill": {
      "post": {
        "tags": [
          "Sync"
        ],
        "operationId": "runOuraBackfill",
        "summary": "Run an Oura backfill sync",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "startDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date"
                  }
                },
                "required": [
                  "startDate",
                  "endDate"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Started and completed an Oura backfill sync.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSyncResponse"
                }
              }
            }
          },
          "400": {
            "description": "The backfill request payload is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Oura reauthentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "An Oura sync is already running.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "The sync failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}